# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rede Akiba API REST",
        "description": "Esta documentação descreve as todas as operações disponíveis para as aplicações da Rede Akiba.",
        "contact": {
            "email": "contato@redeakiba.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/eventos": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Lista todos os eventos cadastrados",
                "description": "Este endpoint retorna todas os cadastrados no sistema da Rede Akiba.",
                "operationId": "6783382f3cfb7184eb56d4084d87ff8b",
                "responses": {
                    "200": {
                        "description": "Lista de todos os eventos cadastrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum evento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Nenhum evento encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Cadastra um novo evento",
                "description": "Este endpoint realiza o cadastro de um novo evento no sistema da Rede Akiba.",
                "operationId": "c978611e66062e994ed94fbd7f47fc42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuário não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ocorreu um erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de validação"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/eventos/{slug}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Retorna um evento específico",
                "description": "Este endpoint retorna um evento específico cadastrado no sistema da Rede Akiba.",
                "operationId": "5e01d4ccc9c9c71d1fe0bb38949e53cc",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug do Evento: Retorna um evento específico baseado no slug fornecido.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Evento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/eventos/{id}": {
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Remove um evento específico",
                "description": "Este endpoint remove um evento específico cadastrado no sistema da Rede Akiba.",
                "operationId": "450c9e239ce529dfd4373283810db37e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Evento: Remove um evento específico baseado no id fornecido.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Evento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Atualiza um evento específico",
                "description": "Este endpoint atualiza um evento específico cadastrado no sistema da Rede Akiba.",
                "operationId": "22941a9c3d37107714834c35e3b26ffa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Evento: Atualiza um evento específico baseado no id fornecido.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Evento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro de processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocorreu um erro de processamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "EventsResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "dates": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EventsRequest": {
                "properties": {
                    "author": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "dates": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}